Resolved bug in Items not showing list

+++ database/migrations/2023_11_22_145724_create_orders_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
-            $table->decimal('total');
+            $table->float('total');
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
             $table->timestamp('returned_at')->nullable();
             $table->timestamps();
